add fault tolerant to get/put: no such machine error

+++ get.pl
@@ -8,6 +8,10 @@ my $ARGC = @ARGV ;
 ($ARGC == 3) or die("$0 {target} {remote} {local}\n") ;
 
 my ($target, $remote, $local) = @ARGV ;
+if ( ! exists $h_host{$target} ){
+	print STDERR "$target: host does not exists\n" ;
+	return -1 ;
+}
 my $home = $h_host{$target}->{"home"} ;
 my $ret = system qq( scp -C -r "$target:$home/$remote" $local ) ;
